Add an explicit cast to avoid a compiler warning
commands.c:169:16: warning: comparison of integers of different signs:
      'unsigned int' and 'int' [-Wsign-compare]
                if (*nlength > r-1)
                    ~~~~~~~~ ^ ~~~


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@5514 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/commands.c
@@ -166,7 +166,7 @@ RESPONSECODE CmdPowerOn(unsigned int reader_index, unsigned int * nlength,
 		}
 
 		DEBUG_INFO_XXD("Data Block: ", tmp, r);
-		if (*nlength > r-1)
+		if ((int)*nlength > r-1)
 			*nlength = r-1;
 		memcpy(buffer, tmp+1, *nlength);
